uploaded training images to data store, wrote dataset and transforms for training

+++ Image2Letter/image2letterAI/main.py
@@ -3,6 +3,7 @@ import config.config as config
 import requests
 from pathlib import Path
 from utils import load_transposed_convolutions
+import cv2
 
 def elt_data():
     """Extract, load and transform our data assets."""
@@ -12,14 +13,17 @@ def elt_data():
     for url_file in url_file_paths:
         links = open(url_file).read().strip().split("\n")
         output_path = Path(config.TRAINING_IMGS_DIR, os.path.basename(url_file)[:-4])
+        os.makedirs(output_path, exist_ok=True)
         total = 0
         for url in links:
             try:
+                p = os.path.sep.join([str(output_path), "{}.jpg".format(
+                    str(total).zfill(8))])
+                if (os.path.exists(p)):
+                    continue
                 # try to download the image
                 r = requests.get(url, timeout=60)
                 # save the image to disk
-                p = os.path.sep.join([output_path, "{}.jpg".format(
-                    str(total).zfill(8))])
                 f = open(p, "wb")
                 f.write(r.content)
                 f.close()
@@ -30,6 +34,18 @@ def elt_data():
             except:
                 print("[INFO] error downloading {}...skipping".format(p))
 
+def remove_bad_images():
+    for folder in os.scandir(config.TRAINING_IMGS_DIR):
+        files :list[str] = os.listdir(folder.path)
+        for file_path in files:
+            try:
+                _ = cv2.imread(os.path.join(folder.path, file_path), cv2.IMREAD_GRAYSCALE)  
+            except:
+                os.remove(file_path)
+                print("errounous image, deleted ", file_path)
+        
+
 # TODO test this function
-# elt_data()
-load_transposed_convolutions(str(config.FONT_PATH), 34, config.TYPEWRITER_CONFIG["letterList"])
+#remove_bad_images()
+#elt_data()
+# load_transposed_convolutions(str(config.FONT_PATH), 34, config.TYPEWRITER_CONFIG["letterList"])
